Remove AbstractIndividual
