+++ gp/hl_api.py
@@ -22,7 +22,7 @@ def evolve(
         A population class that will be evolved.
     objective : Callable
         An objective function used for the evolution. Needs to take an
-        invidual (gp.AbstractIndividual) as input parameter and return
+        invidual (gp.CGPIndividual) as input parameter and return
         a modified individual (with updated fitness).
     ea : EA algorithm instance
         The evolution algorithm. Needs to be a class instance with an
